Clean up button stories
